more tests passing

+++ bootstrap/include/frost/io/ByteOrder.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "ByteOrder_types.h"
 #include "frost/io/ByteOrder_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void frost$io$ByteOrder$cleanup(frost$io$ByteOrder self);
-frost$io$ByteOrder frost$io$ByteOrder$init$frost$core$Int64(frost$core$Int64 p_rv);
+frost$io$ByteOrder frost$io$ByteOrder$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit frost$io$ByteOrder$$EQ$frost$io$ByteOrder$R$frost$core$Bit(frost$io$ByteOrder self, frost$io$ByteOrder p_other);
