+++ bootstrap/include/frost/json/JSON.h
@@ -10,6 +10,7 @@ typedef struct frost$json$JSON frost$json$JSON;
 #include "frost/core/Bit_types.h"
 typedef struct frost$collections$ListView frost$collections$ListView;
 typedef struct frost$collections$MapView frost$collections$MapView;
+#include "frost/core/Int_types.h"
 
 frost$json$JSON$PathKey* frost$json$JSON$nextPathKey$frost$core$String$frost$core$String$Index$R$frost$json$JSON$PathKey$Q(frost$core$String* p_path, frost$core$String$Index p_start);
 frost$json$JSON* frost$json$JSON$get$frost$core$String$R$frost$json$JSON$Q(frost$json$JSON* self, frost$core$String* p_path);
@@ -26,11 +27,11 @@ frost$collections$MapView* frost$json$JSON$get_asMap$R$frost$collections$MapView
 frost$collections$ListView* frost$json$JSON$get_asListView$R$frost$collections$ListView$LTfrost$json$JSON$GT(frost$json$JSON* self);
 frost$core$String* frost$json$JSON$get_asString$R$frost$core$String(frost$json$JSON* self);
 void frost$json$JSON$cleanup(frost$json$JSON* self);
-void frost$json$JSON$init$frost$core$Int64$frost$core$Int64(frost$json$JSON* self, frost$core$Int64 p_rv, frost$core$Int64 p_f0);
-void frost$json$JSON$init$frost$core$Int64$frost$core$Real64(frost$json$JSON* self, frost$core$Int64 p_rv, frost$core$Real64 p_f0);
-void frost$json$JSON$init$frost$core$Int64$frost$core$String(frost$json$JSON* self, frost$core$Int64 p_rv, frost$core$String* p_f0);
-void frost$json$JSON$init$frost$core$Int64$frost$core$Bit(frost$json$JSON* self, frost$core$Int64 p_rv, frost$core$Bit p_f0);
-void frost$json$JSON$init$frost$core$Int64$frost$collections$MapView$LTfrost$core$String$Cfrost$json$JSON$GT(frost$json$JSON* self, frost$core$Int64 p_rv, frost$collections$MapView* p_f0);
-void frost$json$JSON$init$frost$core$Int64$frost$collections$ListView$LTfrost$json$JSON$GT(frost$json$JSON* self, frost$core$Int64 p_rv, frost$collections$ListView* p_f0);
-void frost$json$JSON$init$frost$core$Int64(frost$json$JSON* self, frost$core$Int64 p_rv);
+void frost$json$JSON$init$frost$core$Int$frost$core$Int64(frost$json$JSON* self, frost$core$Int p_rv, frost$core$Int64 p_f0);
+void frost$json$JSON$init$frost$core$Int$frost$core$Real64(frost$json$JSON* self, frost$core$Int p_rv, frost$core$Real64 p_f0);
+void frost$json$JSON$init$frost$core$Int$frost$core$String(frost$json$JSON* self, frost$core$Int p_rv, frost$core$String* p_f0);
+void frost$json$JSON$init$frost$core$Int$frost$core$Bit(frost$json$JSON* self, frost$core$Int p_rv, frost$core$Bit p_f0);
+void frost$json$JSON$init$frost$core$Int$frost$collections$MapView$LTfrost$core$String$Cfrost$json$JSON$GT(frost$json$JSON* self, frost$core$Int p_rv, frost$collections$MapView* p_f0);
+void frost$json$JSON$init$frost$core$Int$frost$collections$ListView$LTfrost$json$JSON$GT(frost$json$JSON* self, frost$core$Int p_rv, frost$collections$ListView* p_f0);
+void frost$json$JSON$init$frost$core$Int(frost$json$JSON* self, frost$core$Int p_rv);
