+++ bootstrap/include/org/frostlang/frostc/ClassDecl/Kind.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "Kind_types.h"
 #include "org/frostlang/frostc/ClassDecl/Kind_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void org$frostlang$frostc$ClassDecl$Kind$cleanup(org$frostlang$frostc$ClassDecl$Kind self);
-org$frostlang$frostc$ClassDecl$Kind org$frostlang$frostc$ClassDecl$Kind$init$frost$core$Int64(frost$core$Int64 p_rv);
+org$frostlang$frostc$ClassDecl$Kind org$frostlang$frostc$ClassDecl$Kind$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit org$frostlang$frostc$ClassDecl$Kind$$EQ$org$frostlang$frostc$ClassDecl$Kind$R$frost$core$Bit(org$frostlang$frostc$ClassDecl$Kind self, org$frostlang$frostc$ClassDecl$Kind p_other);
