+++ bootstrap/include/org/frostlang/frostc/Compiler/TypeContext.h
@@ -3,14 +3,14 @@
 #include "TypeContext_types.h"
 typedef struct frost$core$String frost$core$String;
 typedef struct org$frostlang$frostc$Compiler$TypeContext org$frostlang$frostc$Compiler$TypeContext;
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 typedef struct org$frostlang$frostc$Type org$frostlang$frostc$Type;
 #include "frost/core/Bit_types.h"
 
 frost$core$String* org$frostlang$frostc$Compiler$TypeContext$description$R$frost$core$String(org$frostlang$frostc$Compiler$TypeContext* self);
 frost$core$String* org$frostlang$frostc$Compiler$TypeContext$get_asString$R$frost$core$String(org$frostlang$frostc$Compiler$TypeContext* self);
 void org$frostlang$frostc$Compiler$TypeContext$cleanup(org$frostlang$frostc$Compiler$TypeContext* self);
-void org$frostlang$frostc$Compiler$TypeContext$init$frost$core$Int64(org$frostlang$frostc$Compiler$TypeContext* self, frost$core$Int64 p_rv);
-void org$frostlang$frostc$Compiler$TypeContext$init$frost$core$Int64$org$frostlang$frostc$Type$frost$core$Bit(org$frostlang$frostc$Compiler$TypeContext* self, frost$core$Int64 p_rv, org$frostlang$frostc$Type* p_f0, frost$core$Bit p_f1);
+void org$frostlang$frostc$Compiler$TypeContext$init$frost$core$Int(org$frostlang$frostc$Compiler$TypeContext* self, frost$core$Int p_rv);
+void org$frostlang$frostc$Compiler$TypeContext$init$frost$core$Int$org$frostlang$frostc$Type$frost$core$Bit(org$frostlang$frostc$Compiler$TypeContext* self, frost$core$Int p_rv, org$frostlang$frostc$Type* p_f0, frost$core$Bit p_f1);
 frost$core$Bit org$frostlang$frostc$Compiler$TypeContext$$EQ$org$frostlang$frostc$Compiler$TypeContext$R$frost$core$Bit(org$frostlang$frostc$Compiler$TypeContext* self, org$frostlang$frostc$Compiler$TypeContext* p_other);
