+++ bootstrap/include/org/frostlang/frostc/FieldDecl/Kind.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "Kind_types.h"
 #include "org/frostlang/frostc/FieldDecl/Kind_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void org$frostlang$frostc$FieldDecl$Kind$cleanup(org$frostlang$frostc$FieldDecl$Kind self);
-org$frostlang$frostc$FieldDecl$Kind org$frostlang$frostc$FieldDecl$Kind$init$frost$core$Int64(frost$core$Int64 p_rv);
+org$frostlang$frostc$FieldDecl$Kind org$frostlang$frostc$FieldDecl$Kind$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit org$frostlang$frostc$FieldDecl$Kind$$EQ$org$frostlang$frostc$FieldDecl$Kind$R$frost$core$Bit(org$frostlang$frostc$FieldDecl$Kind self, org$frostlang$frostc$FieldDecl$Kind p_other);
