+++ bootstrap/include/org/frostlang/frostc/IR/Value_types.h
@@ -3,14 +3,14 @@
 typedef struct frost$core$Class frost$core$Class;
 #include "frost/core/Int32_types.h"
 #include "frost/core/UInt8_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 typedef struct frost$core$String frost$core$String;
 
 typedef struct org$frostlang$frostc$IR$Value {
     frost$core$Class* $class;
     frost$core$Int32 refCount;
     frost$core$UInt8 $flags;
-    frost$core$Int64 $rawValue;
+    frost$core$Int $rawValue;
     int8_t $data[24];
 } org$frostlang$frostc$IR$Value;
 #include "frost/core/Class.h"
