+++ bootstrap/include/org/frostlang/frostc/LLVMCodeGenerator/OpClass_types.h
@@ -1,12 +1,12 @@
 #pragma once
 #include "frost_c.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 typedef struct frost$core$String frost$core$String;
 typedef struct frost$core$Class frost$core$Class;
 #include "org/frostlang/frostc/LLVMCodeGenerator/OpClass_types.h"
 
 typedef struct org$frostlang$frostc$LLVMCodeGenerator$OpClass {
-    frost$core$Int64 $rawValue;
+    frost$core$Int $rawValue;
     int8_t $data[0];
 } org$frostlang$frostc$LLVMCodeGenerator$OpClass;
 #include "frost/core/Class.h"
