+++ bootstrap/include/org/frostlang/frostc/Main/Format.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "Format_types.h"
 #include "org/frostlang/frostc/Main/Format_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void org$frostlang$frostc$Main$Format$cleanup(org$frostlang$frostc$Main$Format self);
-org$frostlang$frostc$Main$Format org$frostlang$frostc$Main$Format$init$frost$core$Int64(frost$core$Int64 p_rv);
+org$frostlang$frostc$Main$Format org$frostlang$frostc$Main$Format$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit org$frostlang$frostc$Main$Format$$EQ$org$frostlang$frostc$Main$Format$R$frost$core$Bit(org$frostlang$frostc$Main$Format self, org$frostlang$frostc$Main$Format p_other);
