+++ bootstrap/include/org/frostlang/frostc/MethodDecl/Kind.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "Kind_types.h"
 #include "org/frostlang/frostc/MethodDecl/Kind_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void org$frostlang$frostc$MethodDecl$Kind$cleanup(org$frostlang$frostc$MethodDecl$Kind self);
-org$frostlang$frostc$MethodDecl$Kind org$frostlang$frostc$MethodDecl$Kind$init$frost$core$Int64(frost$core$Int64 p_rv);
+org$frostlang$frostc$MethodDecl$Kind org$frostlang$frostc$MethodDecl$Kind$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit org$frostlang$frostc$MethodDecl$Kind$$EQ$org$frostlang$frostc$MethodDecl$Kind$R$frost$core$Bit(org$frostlang$frostc$MethodDecl$Kind self, org$frostlang$frostc$MethodDecl$Kind p_other);
