+++ bootstrap/include/org/frostlang/frostc/Symbol/Kind.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "Kind_types.h"
 #include "org/frostlang/frostc/Symbol/Kind_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void org$frostlang$frostc$Symbol$Kind$cleanup(org$frostlang$frostc$Symbol$Kind self);
-org$frostlang$frostc$Symbol$Kind org$frostlang$frostc$Symbol$Kind$init$frost$core$Int64(frost$core$Int64 p_rv);
+org$frostlang$frostc$Symbol$Kind org$frostlang$frostc$Symbol$Kind$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit org$frostlang$frostc$Symbol$Kind$$EQ$org$frostlang$frostc$Symbol$Kind$R$frost$core$Bit(org$frostlang$frostc$Symbol$Kind self, org$frostlang$frostc$Symbol$Kind p_other);
