+++ bootstrap/include/org/frostlang/frostc/Type/Kind.h
@@ -2,10 +2,10 @@
 #include "frost_c.h"
 #include "Kind_types.h"
 #include "org/frostlang/frostc/Type/Kind_types.h"
-#include "frost/core/Int64_types.h"
+#include "frost/core/Int_types.h"
 #include "frost/core/Bit_types.h"
 
 void org$frostlang$frostc$Type$Kind$cleanup(org$frostlang$frostc$Type$Kind self);
-org$frostlang$frostc$Type$Kind org$frostlang$frostc$Type$Kind$init$frost$core$Int64(frost$core$Int64 p_rv);
+org$frostlang$frostc$Type$Kind org$frostlang$frostc$Type$Kind$init$frost$core$Int(frost$core$Int p_rv);
 frost$core$Bit org$frostlang$frostc$Type$Kind$$EQ$org$frostlang$frostc$Type$Kind$R$frost$core$Bit(org$frostlang$frostc$Type$Kind self, org$frostlang$frostc$Type$Kind p_other);
